move trait into own submodule
