+++ src/wire/mod.rs
@@ -1,5 +1,6 @@
 pub mod marshal;
 pub mod marshal_base;
+pub mod marshal_trait;
 pub mod marshal_container;
 pub mod unmarshal;
 pub mod unmarshal_base;
